Fix format specifiers and enforce type safety in reference_model

+++ reference_model/src/ops/control_flow.cpp
@@ -66,9 +66,9 @@ int OpControlFlow::evalBlock(TosaSerializationBasicBlock* block,
                block_outputs.size(), num_output_tensors);
 
     // set graph traverser's input = basic block's input
-    for (int i = 0; i < num_input_tensors; i++)
+    for (size_t i = 0; i < num_input_tensors; i++)
     {
-        TosaReference::Tensor* tensor = block_sgt.getInputTensor(i);
+        TosaReference::Tensor* tensor = block_sgt.getInputTensor(static_cast<uint32_t>(i));
         ERROR_IF(!tensor->is_allocated(), "block %s input tensor %s are not initialized before use", block_name.c_str(),
                  tensor->getName().c_str());
 
@@ -119,7 +119,7 @@ int OpControlFlow::evalBlock(TosaSerializationBasicBlock* block,
 
     // make sure output tensor is evaluated and show its value
     bool all_output_valid = true;
-    for (int i = 0; i < num_output_tensors; i++)
+    for (uint32_t i = 0; i < num_output_tensors; i++)
     {
         const TosaReference::Tensor* ct = block_sgt.getOutputTensor(i);
         ASSERT_MEM(ct);
@@ -141,9 +141,9 @@ int OpControlFlow::evalBlock(TosaSerializationBasicBlock* block,
     }
 
     // set basic block's output = subgraph_traverser's output
-    for (int i = 0; i < num_output_tensors; i++)
+    for (size_t i = 0; i < num_output_tensors; i++)
     {
-        TosaReference::Tensor* tensor = block_sgt.getOutputTensor(i);
+        TosaReference::Tensor* tensor = block_sgt.getOutputTensor(static_cast<uint32_t>(i));
         ERROR_IF(!tensor->is_allocated(), "block %s input tensor %s are not initialized before use", block_name.c_str(),
                  tensor->getName().c_str());
 
@@ -175,7 +175,7 @@ int OpCondIf<Rank>::checkTensorAttributes()
 {
     ERROR_IF(!tsh, "OpCondIf: tosa serialization handler must not be null");
 
-    int32_t num_inputs = getInputs().size();
+    size_t num_inputs = getInputs().size();
     ERROR_IF(num_inputs < 1, "OpCondIf: must have at least 1 operand");
 
     ERROR_IF(inputs[0]->getDtype() != TOSA_REF_TYPE_BOOL || inputs[0]->getElementCount() != 1,
@@ -212,8 +212,8 @@ int OpCondIf<Rank>::checkTensorAttributes()
 
     // Make sure operator input/output matches block input/output
     // Skip the first rank 0 bool tensor on input list
-    int32_t num_input_tensor  = getInputs().size() - 1;
-    int32_t num_output_tensor = getOutputs().size();
+    size_t num_input_tensor  = getInputs().size() - 1;
+    size_t num_output_tensor = getOutputs().size();
 
     ERROR_IF((int32_t)then_block->GetInputs().size() != num_input_tensor,
              "OpCondIf: then_block has unexpected number of input");
@@ -224,7 +224,7 @@ int OpCondIf<Rank>::checkTensorAttributes()
     ERROR_IF((int32_t)else_block->GetOutputs().size() != num_output_tensor,
              "OpCondIf: else_block has unexpected number of output");
 
-    for (int32_t i = 0; i < num_input_tensor; i++)
+    for (size_t i = 0; i < num_input_tensor; i++)
     {
         Tensor* operator_input                    = getInputs()[i + 1];
         std::string then_block_input_name         = then_block->GetInputs()[i];
@@ -239,7 +239,7 @@ int OpCondIf<Rank>::checkTensorAttributes()
                  "OpCondIf: input tensor rank mismatch with then_block input rank");
         ERROR_IF(operator_input->getRank() != (int32_t)else_block_input->GetShape().size(),
                  "OpCondIf: input tensor rank mismatch with else_block input rank");
-        for (int32_t d = 0; d < operator_input->getRank(); d++)
+        for (size_t d = 0; d < operator_input->getRank(); d++)
         {
             ERROR_IF(operator_input->getShape()[d] != then_block_input->GetShape()[d],
                      "OpCondIf: input tensor dimension mismatch with then_block input dimension");
@@ -248,7 +248,7 @@ int OpCondIf<Rank>::checkTensorAttributes()
         }
     }
 
-    for (int32_t i = 0; i < num_output_tensor; i++)
+    for (size_t i = 0; i < num_output_tensor; i++)
     {
         Tensor* operator_output                    = getOutputs()[i];
         std::string then_block_output_name         = then_block->GetOutputs()[i];
@@ -263,7 +263,7 @@ int OpCondIf<Rank>::checkTensorAttributes()
                  "OpCondIf: output tensor rank mismatch with then_block output rank");
         ERROR_IF(operator_output->getRank() != (int32_t)else_block_output->GetShape().size(),
                  "OpCondIf: output tensor rank mismatch with else_block output rank");
-        for (int32_t d = 0; d < operator_output->getRank(); d++)
+        for (size_t d = 0; d < operator_output->getRank(); d++)
         {
             ERROR_IF(operator_output->getShape()[d] != then_block_output->GetShape()[d],
                      "OpCondIf: output tensor dimension mismatch with then_block output dimension");
@@ -321,7 +321,7 @@ int OpWhileLoop::checkTensorAttributes()
         return 1;
     }
 
-    int32_t num_inputs = getInputs().size();
+    size_t num_inputs = getInputs().size();
     if (num_inputs <= 0)
     {
         WARNING("OpWhileLoop: must have at least 1 operands");
@@ -358,7 +358,7 @@ int OpWhileLoop::checkTensorAttributes()
     ERROR_IF(!body_block, "OpWhileLoop: fail to resolve body_graph %s", attribute->body_graph().c_str());
 
     // Make sure operator input/output matches block input/output
-    int32_t num_block_tensor = getInputs().size();
+    size_t num_block_tensor = getInputs().size();
     ERROR_IF((int32_t)getOutputs().size() != num_block_tensor,
              "OpWhileLoop: operator input tensor doesn't match output");
     ERROR_IF((int32_t)cond_block->GetInputs().size() != num_block_tensor,
@@ -367,7 +367,7 @@ int OpWhileLoop::checkTensorAttributes()
              "OpWhileLoop: body_block has unexpected number of input");
     ERROR_IF((int32_t)body_block->GetOutputs().size() != num_block_tensor,
              "OpWhileLoop: body_block has unexpected number of output");
-    for (int32_t i = 0; i < num_block_tensor; i++)
+    for (size_t i = 0; i < num_block_tensor; i++)
     {
         Tensor* operator_input  = getInputs()[i];
         Tensor* operator_output = getOutputs()[i];
@@ -394,7 +394,7 @@ int OpWhileLoop::checkTensorAttributes()
         ERROR_IF(operator_input->getRank() != (int32_t)body_block_output->GetShape().size(),
                  "OpWhileLoop: input tensor rank mismatch with body_block output rank");
 
-        for (int32_t d = 0; d < operator_input->getRank(); d++)
+        for (size_t d = 0; d < operator_input->getRank(); d++)
         {
             ERROR_IF(operator_input->getShape()[d] != cond_block_input->GetShape()[d],
                      "OpWhileLoop: input tensor dimension mismatch with cond_block input dimension");
