+++ reference_model/src/ops/custom.cpp
@@ -1,5 +1,5 @@
 
-// Copyright (c) 2020, 2023-2024, ARM Limited.
+// Copyright (c) 2020, 2023-2025, ARM Limited.
 //
 //    Licensed under the Apache License, Version 2.0 (the "License");
 //    you may not use this file except in compliance with the License.
@@ -49,9 +49,9 @@ int OpCustom::checkTensorAttributes()
 
 int OpCustom::eval()
 {
-    auto inputs        = getInputs();
-    int32_t num_inputs = inputs.size();
-    auto tosa_level    = g_func_config.tosa_level;
+    auto inputs       = getInputs();
+    size_t num_inputs = inputs.size();
+    auto tosa_level   = g_func_config.tosa_level;
     LEVEL_CHECK(num_inputs <= tosa_level.MAX_TENSOR_LIST_SIZE,
                 "num_inputs should be smaller than or equal to MAX_TENSOR_LIST_SIZE");
